[BREAKING] Opt(Restore): Optimize Restore's new map-reduce based design (#7666)
This PR along with the previous restore PR is a BREAKING change. Marking this PR as breaking, because we forgot to mark the previous one.

- Make restore map run concurrently for faster execution.
- Add progress updates every second for both map and reduce phase.
- Refactor code to break out the map and reduce code into separate files.
- Make reduce cheap by avoiding marshal-unmarshal step.

With these changes, I see map phase is faster than reduce and both finish in about 2 mins each. Map runs at 200 MBps, while Reduce runs at 130 MBps, processing 20GB of uncompressed data in under 5 mins.

Changes:
* Work on optimizing restore
* Some file moves
* Moved map output to a temp directory.
* Add restoreTs in export-backup

(cherry picked from commit 1c7d4493b494f55b8e271d4c2654f6c63b4e8763)
