+++ systest/backup/common/utils.go
@@ -46,7 +46,7 @@ func RunFailingRestore(t *testing.T, backupLocation, lastDir string, commitTs ui
 	// calling restore.
 	require.NoError(t, os.RemoveAll(restoreDir))
 
-	result := worker.RunRestore("./data/restore", backupLocation, lastDir,
+	result := worker.RunOfflineRestore("./data/restore", backupLocation, lastDir,
 		"", options.Snappy, 0)
 	require.Error(t, result.Err)
 	require.Contains(t, result.Err.Error(), "expected a BackupNum value of 1")
